Split sozo ops into its own crate
