+++ bin/sozo/src/commands/migrate.rs
@@ -1,14 +1,20 @@
-use anyhow::{anyhow, Result};
+use anyhow::{anyhow, Context, Result};
 use clap::Args;
 use dojo_lang::compiler::MANIFESTS_DIR;
-use dojo_world::metadata::dojo_metadata_from_workspace;
-use scarb::core::Config;
+use dojo_world::metadata::{dojo_metadata_from_workspace, Environment};
+use scarb::core::{Config, Workspace};
+use sozo_ops::migration;
+use starknet::accounts::{Account, ConnectedAccount, SingleOwnerAccount};
+use starknet::core::types::{BlockId, BlockTag, FieldElement, StarknetError};
+use starknet::core::utils::parse_cairo_short_string;
+use starknet::providers::jsonrpc::HttpTransport;
+use starknet::providers::{JsonRpcClient, Provider, ProviderError};
+use starknet::signers::LocalWallet;
 
 use super::options::account::AccountOptions;
 use super::options::starknet::StarknetOptions;
 use super::options::transaction::TransactionOptions;
 use super::options::world::WorldOptions;
-use crate::ops::migration;
 
 #[derive(Args)]
 pub struct MigrateArgs {
@@ -35,6 +41,51 @@ pub struct MigrateArgs {
     pub transaction: TransactionOptions,
 }
 
+pub async fn setup_env<'a>(
+    ws: &'a Workspace<'a>,
+    account: AccountOptions,
+    starknet: StarknetOptions,
+    world: WorldOptions,
+    name: Option<&'a String>,
+    env: Option<&'a Environment>,
+) -> Result<(
+    Option<FieldElement>,
+    SingleOwnerAccount<JsonRpcClient<HttpTransport>, LocalWallet>,
+    String,
+)> {
+    let ui = ws.config().ui();
+
+    let world_address = world.address(env).ok();
+
+    let (account, chain_id) = {
+        let provider = starknet.provider(env)?;
+        let chain_id = provider.chain_id().await?;
+        let chain_id = parse_cairo_short_string(&chain_id)
+            .with_context(|| "Cannot parse chain_id as string")?;
+
+        let mut account = account.account(provider, env).await?;
+        account.set_block_id(BlockId::Tag(BlockTag::Pending));
+
+        let address = account.address();
+
+        ui.print(format!("\nMigration account: {address:#x}"));
+        if let Some(name) = name {
+            ui.print(format!("\nWorld name: {name}\n"));
+        }
+
+        match account.provider().get_class_hash_at(BlockId::Tag(BlockTag::Pending), address).await {
+            Ok(_) => Ok((account, chain_id)),
+            Err(ProviderError::StarknetError(StarknetError::ContractNotFound)) => {
+                Err(anyhow!("Account with address {:#x} doesn't exist.", account.address()))
+            }
+            Err(e) => Err(e.into()),
+        }
+    }
+    .with_context(|| "Problem initializing account for migration.")?;
+
+    Ok((world_address, account, chain_id))
+}
+
 impl MigrateArgs {
     pub fn run(mut self, config: &Config) -> Result<()> {
         let ws = scarb::ops::read_workspace(config.manifest_path(), config)?;
@@ -57,7 +108,19 @@ impl MigrateArgs {
             return Err(anyhow!("Build project using `sozo build` first"));
         }
 
-        ws.config().tokio_handle().block_on(migration::execute(&ws, self, env_metadata))?;
+        let _ = config.tokio_handle().block_on(async {
+            let (world_address, account, chain_id) = setup_env(
+                &ws,
+                self.account,
+                self.starknet,
+                self.world,
+                self.name.as_ref(),
+                env_metadata.as_ref(),
+            )
+            .await?;
+
+            migration::migrate(&ws, world_address, chain_id, &account, self.name).await
+        });
 
         Ok(())
     }
