+++ bin/sozo/src/lib.rs
@@ -1,4 +1,3 @@
 pub mod args;
 pub mod commands;
-pub mod ops;
 pub mod utils;
