+++ crates/dojo-world/src/migration/mod.rs
@@ -85,6 +85,8 @@ pub struct TxConfig {
     /// The multiplier for how much the actual transaction max fee should be relative to the
     /// estimated fee. If `None` is provided, the multiplier is set to `1.1`.
     pub fee_estimate_multiplier: Option<f64>,
+    pub wait: bool,
+    pub receipt: bool,
 }
 
 #[cfg_attr(not(target_arch = "wasm32"), async_trait)]
@@ -114,7 +116,7 @@ pub trait Declarable {
 
         let mut txn = account.declare(Arc::new(flattened_class), casm_class_hash);
 
-        if let TxConfig { fee_estimate_multiplier: Some(multiplier) } = txn_config {
+        if let TxConfig { fee_estimate_multiplier: Some(multiplier), .. } = txn_config {
             txn = txn.fee_estimate_multiplier(multiplier);
         }
 
@@ -193,7 +195,7 @@ pub trait Deployable: Declarable + Sync {
 
         let mut txn = account.execute(vec![call]);
 
-        if let TxConfig { fee_estimate_multiplier: Some(multiplier) } = txn_config {
+        if let TxConfig { fee_estimate_multiplier: Some(multiplier), .. } = txn_config {
             txn = txn.fee_estimate_multiplier(multiplier);
         }
 
@@ -258,7 +260,7 @@ pub trait Deployable: Declarable + Sync {
             to: felt!("0x41a78e741e5af2fec34b695679bc6891742439f7afb8484ecd7766661ad02bf"),
         }]);
 
-        if let TxConfig { fee_estimate_multiplier: Some(multiplier) } = txn_config {
+        if let TxConfig { fee_estimate_multiplier: Some(multiplier), .. } = txn_config {
             txn = txn.fee_estimate_multiplier(multiplier);
         }
