+++ crates/sozo/ops/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 edition.workspace = true
-name = "sozo"
+name = "sozo-ops"
 version.workspace = true
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
@@ -44,8 +44,6 @@ tokio.workspace = true
 tracing-log = "0.1.3"
 tracing.workspace = true
 url.workspace = true
-sozo-ops.workspace = true
-
 cainome = { git = "https://github.com/cartridge-gg/cainome", tag = "v0.2.2" }
 
 [dev-dependencies]
