+++ crates/sozo/ops/src/auth.rs
@@ -1,0 +1,172 @@
+use std::str::FromStr;
+
+use anyhow::{Context, Result};
+use dojo_world::contracts::model::ModelError;
+use dojo_world::contracts::world::WorldContract;
+use dojo_world::contracts::{cairo_utils, WorldContractReader};
+use dojo_world::migration::TxConfig;
+use starknet::accounts::ConnectedAccount;
+use starknet::core::utils::parse_cairo_short_string;
+use starknet::providers::Provider;
+use starknet_crypto::FieldElement;
+
+use super::get_contract_address;
+use crate::utils::handle_transaction_result;
+
+#[derive(Debug, Clone, PartialEq)]
+pub enum ResourceType {
+    Contract(String),
+    Model(FieldElement),
+}
+
+#[derive(Debug, Clone, PartialEq)]
+pub struct ModelContract {
+    pub model: FieldElement,
+    pub contract: String,
+}
+
+impl FromStr for ModelContract {
+    type Err = anyhow::Error;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        let parts: Vec<&str> = s.split(',').collect();
+
+        let (model, contract) = match parts.as_slice() {
+            [model, contract] => (model, contract),
+            _ => anyhow::bail!(
+                "Model and contract address are expected to be comma separated: `sozo auth writer \
+                 model_name,0x1234`"
+            ),
+        };
+
+        let model = cairo_utils::str_to_felt(model)
+            .map_err(|_| anyhow::anyhow!("Invalid model name: {}", model))?;
+
+        Ok(ModelContract { model, contract: contract.to_string() })
+    }
+}
+
+#[derive(Debug, Clone, PartialEq)]
+pub struct OwnerResource {
+    pub resource: ResourceType,
+    pub owner: FieldElement,
+}
+
+impl FromStr for OwnerResource {
+    type Err = anyhow::Error;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        let parts: Vec<&str> = s.split(',').collect();
+
+        let (resource_part, owner_part) = match parts.as_slice() {
+            [resource, owner] => (*resource, *owner),
+            _ => anyhow::bail!(
+                "Owner and resource are expected to be comma separated: `sozo auth owner \
+                 resource_type:resource_name,0x1234`"
+            ),
+        };
+
+        let owner = FieldElement::from_hex_be(owner_part)
+            .map_err(|_| anyhow::anyhow!("Invalid owner address: {}", owner_part))?;
+
+        let resource_parts = resource_part.split_once(':');
+        let resource = match resource_parts {
+            Some(("contract", name)) => ResourceType::Contract(name.to_string()),
+            Some(("model", name)) => {
+                let model = cairo_utils::str_to_felt(name)
+                    .map_err(|_| anyhow::anyhow!("Invalid model name: {}", name))?;
+                ResourceType::Model(model)
+            }
+            _ => anyhow::bail!(
+                "Resource is expected to be in the format `resource_type:resource_name`: `sozo \
+                 auth owner 0x1234,resource_type:resource_name`"
+            ),
+        };
+
+        Ok(OwnerResource { owner, resource })
+    }
+}
+
+pub async fn grant_writer<A, P>(
+    world: &WorldContract<A>,
+    models_contracts: Vec<ModelContract>,
+    world_reader: WorldContractReader<P>,
+    transaction: TxConfig,
+) -> Result<()>
+where
+    A: ConnectedAccount + Sync + Send + 'static,
+    P: Provider + Sync + Send,
+{
+    let mut calls = Vec::new();
+
+    for mc in models_contracts {
+        let model_name = parse_cairo_short_string(&mc.model)?;
+        match world_reader.model_reader(&model_name).await {
+            Ok(_) => {
+                let contract = get_contract_address(world, mc.contract).await?;
+                calls.push(world.grant_writer_getcall(&mc.model, &contract.into()));
+            }
+
+            Err(ModelError::ModelNotFound) => {
+                println!("Unknown model '{}' => IGNORED", model_name);
+            }
+
+            Err(err) => {
+                return Err(err.into());
+            }
+        }
+    }
+
+    if !calls.is_empty() {
+        let res = world
+            .account
+            .execute(calls)
+            .send()
+            .await
+            .with_context(|| "Failed to send transaction")?;
+
+        handle_transaction_result(
+            &world.account.provider(),
+            res,
+            transaction.wait,
+            transaction.receipt,
+        )
+        .await?;
+    }
+    Ok(())
+}
+
+pub async fn grant_owner<A>(
+    world: WorldContract<A>,
+    owners_resources: Vec<OwnerResource>,
+    transaction: TxConfig,
+) -> Result<()>
+where
+    A: ConnectedAccount + Sync + Send + 'static,
+{
+    let mut calls = Vec::new();
+
+    for or in owners_resources {
+        let resource = match &or.resource {
+            ResourceType::Model(name) => *name,
+            ResourceType::Contract(name_or_address) => {
+                get_contract_address(&world, name_or_address.clone()).await?
+            }
+        };
+
+        calls.push(world.grant_owner_getcall(&or.owner.into(), &resource));
+    }
+
+    let res =
+        world.account.execute(calls).send().await.with_context(|| "Failed to send transaction")?;
+
+    handle_transaction_result(
+        &world.account.provider(),
+        res,
+        transaction.wait,
+        transaction.receipt,
+    )
+    .await?;
+
+    Ok(())
+}
