+++ crates/torii/core/Cargo.toml
@@ -35,6 +35,7 @@ tokio = { version = "1.32.0", features = [ "sync" ], default-features = true }
 tokio-stream = "0.1.11"
 tokio-util = "0.7.7"
 tracing.workspace = true
+sozo-ops.workspace = true
 
 [dev-dependencies]
 camino.workspace = true
