+++ crates/torii/core/src/sql_test.rs
@@ -8,7 +8,7 @@ use dojo_test_utils::sequencer::{
 use dojo_world::contracts::world::WorldContractReader;
 use dojo_world::migration::strategy::MigrationStrategy;
 use scarb::ops;
-use sozo::ops::migration::execute_strategy;
+use sozo_ops::migration::execute_strategy;
 use sqlx::sqlite::{SqliteConnectOptions, SqlitePoolOptions};
 use starknet::core::types::{BlockId, BlockTag, Event, FieldElement};
 use starknet::providers::jsonrpc::HttpTransport;
